feat: create workspace module
