+++ src/lib.rs
@@ -1,2 +1,3 @@
 pub mod models;
 pub mod commands;
+pub mod workspace;
